refactor(renderer): move path helper function to helper forlder
