+++ apps/renderer/src/components/StatusBar.tsx
@@ -1,5 +1,5 @@
 import { StatusBarProps } from "../types/component-types";
-import { basename } from "../utils/path-helper";
+import { basename } from "../utils/helpers/path-helper";
 
 export function StatusBar({ filePath, theme, fontSize }: StatusBarProps) {
   const zoomPercent = Math.round((fontSize / 16) * 100);
