+++ apps/renderer/src/utils/helpers/path-helper.ts
